Project split into multiple other projects, added a nice window open animation
