+++ FindMyGames.UI/Controls/ThemedWindow.cs
