+++ FindMyGames.UI/Themes/Generic.xaml
@@ -2,7 +2,9 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:system="clr-namespace:System;assembly=mscorlib"
-    xmlns:controls="clr-namespace:TheMvvmGuys.FindMyGames.UI.Controls">
+    xmlns:controls="clr-namespace:TheMvvmGuys.FindMyGames.UI.Controls"
+    xmlns:Interactions="http://schemas.microsoft.com/xaml/behaviors"
+    xmlns:f="urn:find-my-games">
     <Style TargetType="Button" x:Key="TitleBarButtonStyle">
         <Setter Property="FontFamily" Value="/FindMyGames;component/Fonts/#Segoe MDL2 Assets"/>
         <Setter Property="BorderBrush" Value="Transparent"/>
@@ -53,7 +55,23 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type controls:ThemedWindow}">
-                    <Grid>
+                    <Grid x:Name="Grid">
+                        <Grid.Resources>
+                            <Duration x:Key="OpenDuration">0:0:0.2</Duration>
+                            <Storyboard x:Key="ThemedWindowOpenStoryboard" Target="{Binding RelativeSource={f:FindAncestor controls:ThemedWindow}}">
+                                <ColorAnimation  Storyboard.TargetProperty="Background.(SolidColorBrush.Color)" From="Transparent" To="{DynamicResource WindowBackgroundColor}" Duration="{StaticResource OpenDuration}"/>
+                                <DoubleAnimation Storyboard.TargetProperty="Opacity" From="0" To="1" Duration="{StaticResource OpenDuration}"/>
+                                <DoubleAnimation Storyboard.TargetProperty="RenderTransform.(ScaleTransform.ScaleX)" From="1.1" To="1"  Duration="{StaticResource OpenDuration}"/>
+                                <DoubleAnimation Storyboard.TargetProperty="RenderTransform.(ScaleTransform.ScaleY)" From="1.1" To="1"  Duration="{StaticResource OpenDuration}"/>
+                            </Storyboard>
+                        </Grid.Resources>
+                        <Interactions:Interaction.Triggers>
+                            <Interactions:EventTrigger EventName="Loaded">
+                                <Interactions:ControlStoryboardAction ControlStoryboardOption="Play" 
+                                                                      Storyboard="{StaticResource ThemedWindowOpenStoryboard}" IsEnabled="True"
+                                                                      />
+                            </Interactions:EventTrigger>
+                        </Interactions:Interaction.Triggers>
                         <AdornerDecorator x:Name="Adorner">
                             <Grid Background="{TemplateBinding Background}">
                                 <Grid.RowDefinitions>
@@ -122,5 +140,11 @@
         <Setter Property="Foreground" Value="{DynamicResource WindowForegroundBrush}"/>
         <Setter Property="BorderThickness" Value="1"/>
         <Setter Property="BorderBrush" Value="{DynamicResource ControlBorderBrush}"/>
+        <Setter Property="RenderTransform">
+            <Setter.Value>
+                <ScaleTransform/>
+            </Setter.Value>
+        </Setter>
+        <Setter Property="RenderTransformOrigin" Value="0.5, 0.5"/>
     </Style>
 </ResourceDictionary>
