+++ FindMyGames.UI/Themes/Resources/DarkThemeColors.xaml
