+++ FindMyGames.UI/Themes/ThemeColorEntryCollection.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.ObjectModel;
+
+namespace TheMvvmGuys.FindMyGames.UI.Themes
+{
+    public class ThemeColorEntryCollection : Collection<ThemeColorEntry>
+    {
+        
+    }
+}
