+++ FindMyGames.UI/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ControlzEx" version="4.0.0-alpha0222" targetFramework="net462" />
+  <package id="Microsoft.Xaml.Behaviors.Wpf" version="1.0.1" targetFramework="net462" />
+</packages>
