+++ FindMyGames.Utilities/FindMyGames.Utilities.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>FindMyGames.Utilities</AssemblyName>
+    <RootNamespace>TheMvvmGuys.FindMyGames.Utilities</RootNamespace>
+  </PropertyGroup>
+
+</Project>
