+++ FindMyGames.sln
@@ -17,6 +17,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ExternalDLLs", "ExternalDLL
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FindMyGames.DataBinding", "FindMyGames.DataBinding\FindMyGames.DataBinding.csproj", "{4A99CABB-BF09-4DC1-84F0-1F05E5EA813B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FindMyGames.UI", "FindMyGames.UI\FindMyGames.UI.csproj", "{64B09EFF-7B7F-4474-82EE-79E5CF3ACE39}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FindMyGames.Utilities", "FindMyGames.Utilities\FindMyGames.Utilities.csproj", "{895421C7-C20E-4B9D-9A27-2D5B2CA3D63B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -43,6 +47,14 @@ Global
 		{4A99CABB-BF09-4DC1-84F0-1F05E5EA813B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4A99CABB-BF09-4DC1-84F0-1F05E5EA813B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4A99CABB-BF09-4DC1-84F0-1F05E5EA813B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64B09EFF-7B7F-4474-82EE-79E5CF3ACE39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64B09EFF-7B7F-4474-82EE-79E5CF3ACE39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64B09EFF-7B7F-4474-82EE-79E5CF3ACE39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64B09EFF-7B7F-4474-82EE-79E5CF3ACE39}.Release|Any CPU.Build.0 = Release|Any CPU
+		{895421C7-C20E-4B9D-9A27-2D5B2CA3D63B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{895421C7-C20E-4B9D-9A27-2D5B2CA3D63B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{895421C7-C20E-4B9D-9A27-2D5B2CA3D63B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{895421C7-C20E-4B9D-9A27-2D5B2CA3D63B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
