+++ FindMyGames/BlendPlayground.xaml
@@ -1,17 +1,11 @@
-﻿<controls:ThemedWindow x:Class="TheMvvmGuys.FindMyGames.BlendPlayground"
+﻿<f:ThemedWindow x:Class="TheMvvmGuys.FindMyGames.BlendPlayground"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:b="http://schemas.microsoft.com/xaml/behaviors"
-        xmlns:ui="clr-namespace:TheMvvmGuys.FindMyGames.UI"
-        xmlns:controls="clr-namespace:TheMvvmGuys.FindMyGames.UI.Controls"
-        xmlns:f="urn:find-my-games" 
+        xmlns:f="urn:find-my-games"
         mc:Ignorable="d"
         Title="BlendPlayground" Height="450" Width="800">
-    <Window.Resources>
-        <f:MathExpressionConverter x:Key="MathConverter"/>
-    </Window.Resources>
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="1*"/>
@@ -33,4 +27,4 @@
         <Slider VerticalAlignment="Top" HorizontalAlignment="Center" Width="250" Minimum="120" Maximum="450" x:Name="Slider" Grid.Row="1"/>
         <TextBlock VerticalAlignment="Center" HorizontalAlignment="Center" Text="{f:MathBinding Value, 'value * 5', ElementName=Slider}" Grid.Row="1"/>
     </Grid>
-</controls:ThemedWindow>
+</f:ThemedWindow>
