--- FindMyGames/Themes/ThemeColorEntryCollection.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
-
-namespace TheMvvmGuys.FindMyGames.Themes
-{
-    public class ThemeColorEntryCollection : Collection<ThemeColorEntry>
-    {
-        
-    }
-}
