+++ FindMyGames/Views/FirstStartup/WelcomePage.xaml
@@ -3,11 +3,8 @@
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:TheMvvmGuys.FindMyGames.Views.FirstStartup"
-      xmlns:root="clr-namespace:TheMvvmGuys.FindMyGames"
-      xmlns:themes="clr-namespace:TheMvvmGuys.FindMyGames.Themes"
-      xmlns:ext="clr-namespace:TheMvvmGuys.FindMyGames.Extensions"
       xmlns:xaml="clr-namespace:TheMvvmGuys.FindMyGames.Extensions.Xaml"
+      xmlns:f="urn:find-my-games"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
       Title="WelcomePage">
@@ -31,7 +28,7 @@
                 <ComboBox ItemsSource="{xaml:AppBinding ThemeColorsDictionary.Themes}"
                           SelectedIndex="{xaml:AppBinding ThemeColorsDictionary.SelectedIndex, Mode=TwoWay}">
                     <ComboBox.ItemTemplate>
-                        <DataTemplate DataType="themes:ThemeColorEntry">
+                        <DataTemplate DataType="f:ThemeColorEntry">
                             <TextBlock Text="{Binding Name, Mode=OneTime}" FontSize="15"/>
                         </DataTemplate>
                     </ComboBox.ItemTemplate>
