+++ FindMyGames/Views/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="TheMvvmGuys.FindMyGames.MainWindow"
+﻿<Window x:Class="TheMvvmGuys.FindMyGames.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
